non-root: enable non-root checkpoint/restore
This commit enables checkpointing and restoring of applications as
non-root.

First goal was to enable checkpoint and restore of the env00 and
pthread00 test case.

This uses the information from opts.uid and opts.cap_eff to skip certain
code paths which do not work as non-root.

The kerndat file is saved in $HOME/.criu.kdat. As $HOME is usually not on
a tmpfs there is now a check for nodename and release to see if the system
has been rebooted since last writing the kerndat cache file and mark it
as stale.

Signed-off-by: Adrian Reber <areber@redhat.com>