Dedup char opts that outlive their uses
SET_CHAR_OPTS() takes an option string and dups it to the specified
member of the opts struct. We don't need to do this for options sourced
from the command line (since argv memory outlives the program).
Now that config file lines are buffered we also don't need to do this
for options that are sourced from config files. Stop doing so in both
cases.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>