Workaround for setgroups permissions issues
Calling setgroups returns EPERM in some environments.
It's not clear why nor why we have any supplementary groups
to begin with. Avoid until root cause is known.

It would be better to guard this with opts.unprivileged,
but it looks like we don't have opts in the restorer, so
use uid instead for the time being.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>