WIP: non-root: Ignore dump uids and gids
Skip restoring process uid and gids in unprivileged mode.
In unprivileged mode "restoring" these ids only works if
the process is already running with the target ids (in which
case the operations are effectively no-ops) or has
the required privileges to change uids/gids. In most
cases we probably want to run as the current uid/gids, which
will either be the same as at checkpoint time, or will be
the desired uid/gids for the current environment.

TODO: This functionality should probably have a corresponding
command line option to enable/disable, in case the user actually
wants to change uid/gids and has the necessary privileges to do so.
Unfortunately this code runs in a context that doesn't have easy
access to command line options.

Signed-off-by: Younes Manton <ymanton@ca.ibm.com>